apply airbnb redux good practices for dashboard and chart container component:
- Only containers should call connect. No connect calls in /Components.
- Containers should not create any new Components inside the file.
