+++ superset/assets/javascripts/chart/Chart.jsx
@@ -5,14 +5,31 @@ import Mustache from 'mustache';
 
 import { d3format } from '../modules/utils';
 import ChartBody from './ChartBody';
+import StackTraceMessage from '../components/StackTraceMessage';
+import visMap from '../../visualizations/main';
 
 const propTypes = {
+  actions: PropTypes.object,
+  chartKey: PropTypes.string.isRequired,
   containerId: PropTypes.string.isRequired,
   datasource: PropTypes.object.isRequired,
   formData: PropTypes.object.isRequired,
   height: PropTypes.number,
   width: PropTypes.number,
-  isLoading: PropTypes.bool,
+  setControlValue: PropTypes.func,
+  timeout: PropTypes.number,
+  viz_type: PropTypes.string.isRequired,
+  // state
+  chartAlert: PropTypes.string,
+  chartStatus: PropTypes.string,
+  chartUpdateEndTime: PropTypes.number,
+  chartUpdateStartTime: PropTypes.number,
+  latestQueryFormData: PropTypes.object,
+  queryRequest: PropTypes.object,
+  queryResponse: PropTypes.object,
+  triggerRender: PropTypes.bool,
+  triggerQuery: PropTypes.bool,
+  // dashboard callbacks
   addFilter: PropTypes.func,
   getFilters: PropTypes.func,
   clearFilter: PropTypes.func,
@@ -41,11 +58,29 @@ class Chart extends React.PureComponent {
     this.removeFilter = this.removeFilter.bind(this);
   }
 
+  componentDidMount() {
+    this.runQuery();
+  }
+
   componentWillReceiveProps(nextProps) {
     this.containerId = nextProps.containerId;
     this.selector = `#${this.containerId}`;
     this.formData = nextProps.formData;
     this.datasource = nextProps.datasource;
+  }
+
+  componentDidUpdate(prevProps) {
+    if (
+      this.props.queryResponse &&
+      this.props.chartStatus === 'success' &&
+      !this.props.queryResponse.error && (
+      prevProps.queryResponse !== this.props.queryResponse ||
+      prevProps.height !== this.props.height ||
+      prevProps.width !== this.props.width ||
+      this.props.triggerRender)
+    ) {
+      this.renderViz();
+    }
   }
 
   getFilters() {
@@ -89,6 +124,13 @@ class Chart extends React.PureComponent {
     return d3format(format, number);
   }
 
+  runQuery() {
+    this.props.actions.runQuery(this.props.formData, true,
+      this.props.timeout,
+      this.props.chartKey,
+    );
+  }
+
   render_template(s) {
     const context = {
       width: this.width(),
@@ -97,23 +139,46 @@ class Chart extends React.PureComponent {
     return Mustache.render(s, context);
   }
 
+  renderViz() {
+    this.props.actions.renderTriggered(this.props.chartKey);
+
+    const viz = visMap[this.props.viz_type];
+    try {
+      viz(this, this.props.queryResponse, this.props.actions.setControlValue);
+    } catch (e) {
+      this.props.actions.chartRenderingFailed(e, this.props.chartKey);
+    }
+  }
+
   render() {
+    const isLoading = this.props.chartStatus === 'loading';
     return (
-      <div className={`chart-body${this.props.isLoading ? ' isloading' : ''}`}>
-        {this.props.isLoading && <img
+      <div className={`token col-md-12 ${isLoading ? 'is-loading' : ''}`}>
+        {isLoading && <img
           alt="loading"
           width="25"
           src="/static/assets/images/loading.gif"
           style={{ position: 'absolute' }}
+        />}
+
+        {this.props.chartAlert &&
+        <StackTraceMessage
+          message={this.props.chartAlert}
+          queryResponse={this.props.queryResponse}
         />
         }
+
+        {!this.props.chartAlert &&
         <ChartBody
           containerId={this.containerId}
           vizType={this.props.formData.viz_type}
           height={this.height.bind(this)}
           width={this.width.bind(this)}
-          ref={(inner) => { this.container = inner; }}
+          ref={(inner) => {
+            this.container = inner;
+          }}
         />
+        }
       </div>
     );
   }
